backend/config/requirements.txt
# Django核心
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0

# 数据库
mysqlclient==2.2.1
pymongo==4.6.1

# 认证
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# 异步任务
celery==5.3.6
redis==5.0.1

# API文档
drf-yasg==1.21.7
coreapi==2.3.3

# NLP和机器学习
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4
gensim==4.3.2
jieba==0.42.1
spacy==3.7.2

# 深度学习（可选，根据需要安装）
# torch==2.1.2
# tensorflow==2.15.0
# transformers==4.36.2

# 爬虫
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.16.0

# 数据处理
lxml==5.1.0
fake-useragent==1.4.0

# 搜索引擎
elasticsearch==8.11.1

# 工具
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# 代码质量
black==23.12.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# 部署
gunicorn==21.2.0
uvicorn==0.25.0

# 其他
Pillow==10.2.0
django-admin-interface==0.28.0
